Run unit tests in CI, type portability fixes
