+++ psx_seq_player/lib_snd.cpp
@@ -79,90 +79,13 @@ extern "C"
 
     LIBVAR unsigned int _snd_vmask;
 
-    struct RegBufStruct
-    {
-        short field_0_vol_left;
-        short field_2_vol_right;
-        short field_4_pitch;
-        unsigned short field_6_vagAddr;
-        unsigned short field_8_adsr1;
-        unsigned short field_A_adsr2;
-        short field_0xc;
-        short field_0xe;
-    };
     LIBVAR RegBufStruct _svm_sreg_buf[24];
     LIBVAR unsigned char _svm_sreg_dirty[24];
 
     LIBVAR int _svm_envx_ptr;
     LIBVAR unsigned int _svm_envx_hist[16];
 
-    struct SpuVoice
-    {
-        short field_0_vag_idx;
-        short field_0x2;
-        short field_4_pitch;
-        short field_6_keyStat;
-        short field_8_voll;
-        char field_0xa;
-        char pad5;
-        short field_C_channel_idx;
-        short field_E_note;
-        short field_10_seq_sep_no;
-        short field_12_fake_program;
-        short field_14_program;
-        short field_16_vag_num; // TODO: char
-        short field_18_vabId;
-        short field_1A_priority;
-
-        char pad8;
-        char field_0x1d;
-
-        short field_1E_bAutoVol;
-        short field_20_autoVolAmount;
-        short field_22_autoVol_dt1;
-        short field_24_autoVol_dt2;
-        short field_26_autoVol_Start;
-        short field_28_autoVol_End;
-        short field_2A_bAutoPan;
-
-        short field_0x2c; // amount ?
-        short field_0x2e; // dt1?
-        short field_0x30; // dt2?
-        short field_0x32; // start ?
-        short pad1;       // end ?
-
-        short field_36_voll;
-    };
-
     LIBVAR SpuVoice _svm_voice[24];
-
-    struct struct_svm
-    {
-        char field_0_sep_sep_no_tonecount;
-        char field_1_vabId;
-        char field_2_note;
-        char field_0x3;
-        char field_4_voll;
-        char field_0x5;
-        char field_6_program;
-        char field_7_fake_program;
-        char field_8_unknown;
-        char field_0x9;
-        char field_A_mvol;
-        char field_B_mpan;
-        char field_C_vag_idx;
-        char field_D_vol;
-        char field_E_pan;
-        char field_F_prior;
-        char field_10_centre;
-        unsigned char field_11_shift;
-        char field_12_mode;
-        char field_0x13;
-        short field_14_seq_sep_no;
-        short field_16_vag_idx;
-        short field_18_voice_idx;
-        int field_0x1a;
-    }; // 26 bytes, can't be bigger than 28 ?
 
     LIBVAR struct_svm _svm_cur;
 
@@ -3472,465 +3395,3 @@ extern "C"
     }
 
 } // extern "C"
-
-static void PcsxReduxExit(short exitCode)
-{
-    printf("Exiting with code %d\n", exitCode);
-    (*(volatile unsigned short *)0x1f802082) = exitCode;
-}
-
-#define ASSERT_EQ(value1, value2)                                                           \
-    if ((value1) != (value2))                                                               \
-    {                                                                                       \
-        printf("%d != %d in %s %s:%d\n", value1, value2, __FUNCTION__, __FILE__, __LINE__); \
-        PcsxReduxExit(1);                                                                   \
-    }
-
-static void Test_SsVmSelectToneAndVag()
-{
-    // Set up
-    _svm_cur.field_0_sep_sep_no_tonecount = 5;
-    _svm_cur.field_7_fake_program = 3;
-    _svm_cur.field_2_note = 4;
-
-    VagAtr testVags[(3 * 16) + 16] = {};
-    testVags[(3 * 16) + 0].min = 4;
-    testVags[(3 * 16) + 0].max = 4;
-    testVags[(3 * 16) + 0].vag = 55;
-
-    testVags[(3 * 16) + 3].min = 5;
-    testVags[(3 * 16) + 3].max = 6;
-    testVags[(3 * 16) + 3].vag = 11;
-
-    testVags[(3 * 16) + 3].min = 2;
-    testVags[(3 * 16) + 3].max = 3;
-    testVags[(3 * 16) + 3].vag = 33;
-
-    testVags[(3 * 16) + 4].min = 4;
-    testVags[(3 * 16) + 4].max = 5;
-    testVags[(3 * 16) + 4].vag = 22;
-
-    _svm_tn = testVags;
-
-    unsigned char vags[256] = {};
-    unsigned char vagNums[256] = {};
-    const short ret = _SsVmSelectToneAndVag(vags, vagNums);
-
-    ASSERT_EQ(ret, 2);
-
-    ASSERT_EQ(0, vags[0]);
-    ASSERT_EQ(4, vags[1]);
-
-    ASSERT_EQ(55, vagNums[0]);
-    ASSERT_EQ(22, vagNums[1]);
-}
-
-static void Test_SsVmDoAllocate()
-{
-    for (int i = 0; i < 24; i++)
-    {
-        SpuVoice blank = {};
-        _svm_voice[i] = blank;
-    }
-
-    _svm_cur.field_18_voice_idx = 3;
-    _svm_cur.field_C_vag_idx = 0;
-    _svm_cur.field_7_fake_program = 0;
-
-    for (int i = 0; i < 16; i++)
-    {
-        _svm_envx_hist[i] = 0xFFFFFFFF;
-    }
-
-    ProgAtr progs[16] = {};
-    _svm_pg = progs;
-    _svm_pg[(_svm_cur.field_16_vag_idx - 1) / 2].reserved3 = 1234;
-    for (int i = 0; i < 24; i++)
-    {
-        _svm_sreg_dirty[i] = 0;
-    }
-
-    VagAtr vags[16 * 5] = {};
-    _svm_tn = vags;
-    VagAtr *pVag = _svm_tn + (_svm_cur.field_7_fake_program * 16) + _svm_cur.field_C_vag_idx;
-    pVag->adsr1 = 4455;
-    pVag->adsr2 = 0xFFe1;
-
-    _svm_damper = 31;
-
-    _SsVmDoAllocate();
-
-    ASSERT_EQ(_svm_voice[_svm_cur.field_18_voice_idx].field_6_keyStat, 0x7fff);
-    for (int i = 0; i < 16; i++)
-    {
-        ASSERT_EQ(_svm_envx_hist[i], 0xFFFFFFF7); // check voice bit is cleared
-    }
-
-    ASSERT_EQ(_svm_sreg_buf[_svm_cur.field_18_voice_idx].field_6_vagAddr, 1234); // check vagAddr is copied as is
-    ASSERT_EQ(_svm_sreg_buf[_svm_cur.field_18_voice_idx].field_8_adsr1, 4455);   // check adsr1 is copied as is
-    ASSERT_EQ(_svm_sreg_buf[_svm_cur.field_18_voice_idx].field_A_adsr2, 0xFFFF);
-
-    ASSERT_EQ(_svm_sreg_dirty[_svm_cur.field_18_voice_idx], 0x30 | 0x8); // check correct regs are set
-}
-
-struct MidiTest
-{
-    int mNoteOnCalls;
-    int mNoteOnNote;
-    int mNoteOnVol;
-
-    int mProgramChangeCalls;
-    int mProgramChangeValue;
-
-    int mControlChangeCalls;
-    int mControlChangeValue;
-
-    int mPitchBendCalls;
-};
-static MidiTest gMidiTest = {};
-
-static void Test_SsNoteOn(short seq_no, short sep_no, unsigned char note, unsigned char voll)
-{
-    gMidiTest.mNoteOnCalls++;
-    gMidiTest.mNoteOnNote = note;
-    gMidiTest.mNoteOnVol = voll;
-}
-
-static void Test_SsSetProgramChange(short seq_no, short sep_no, unsigned char programNum)
-{
-    gMidiTest.mProgramChangeCalls++;
-    gMidiTest.mProgramChangeValue = programNum;
-}
-
-static void Test_ControlChange(short seq_no, short sep_no, unsigned char change)
-{
-    gMidiTest.mControlChangeCalls++;
-    gMidiTest.mControlChangeValue = change;
-}
-
-static void Test_SsSetPitchBend(short seq_no, short sep_no)
-{
-    gMidiTest.mPitchBendCalls++;
-}
-
-static void ClearTestData()
-{
-    MidiTest cleared = {};
-    gMidiTest = cleared;
-}
-
-static int DvConv(int dv)
-{
-    return 2 * ((dv*4) + dv);
-}
-
-static void Test_SsGetSeqData()
-{
-    SsFCALL.noteon = Test_SsNoteOn;
-    SsFCALL.programchange = Test_SsSetProgramChange;
-    SsFCALL.control[0] = Test_ControlChange;
-    SsFCALL.pitchbend = Test_SsSetPitchBend;
-
-    // Note off test
-    {
-        ClearTestData();
-
-        SeqStruct seq = {};
-        _ss_score[0] = &seq;
-
-        unsigned char midiData[] = 
-        {
-            0x95, // Note on, channel 5
-            0xA, // Note
-            0x0, // Velocity (off)
-            0x2, // Delta value
-        };
-        seq.field_0_seq_ptr = midiData;
-
-        _SsGetSeqData(0, 0);
-
-        ASSERT_EQ(seq.field_17_channel_idx, 5);
-        ASSERT_EQ(gMidiTest.mNoteOnCalls, 1);
-        ASSERT_EQ(gMidiTest.mNoteOnNote, 0xA);
-        ASSERT_EQ(gMidiTest.mNoteOnVol, 0x0);
-        ASSERT_EQ(seq.field_16_running_status, 0x90);
-        ASSERT_EQ(seq.field_90_delta_value, DvConv(0x2));
-    }
-
-    // Note off test running status
-    {
-        ClearTestData();
-
-        SeqStruct seq = {};
-        _ss_score[0] = &seq;
-        seq.field_16_running_status = 0x90;
-        seq.field_17_channel_idx = 5;
-        unsigned char midiData[] = 
-        {
-            0xA, // Note
-            0x0, // Velocity (off)
-            0x2, // Delta value
-        };
-        seq.field_0_seq_ptr = midiData;
-
-        _SsGetSeqData(0, 0);
-
-        ASSERT_EQ(seq.field_17_channel_idx, 5);
-        ASSERT_EQ(gMidiTest.mNoteOnCalls, 1);
-        ASSERT_EQ(gMidiTest.mNoteOnNote, 0xA);
-        ASSERT_EQ(gMidiTest.mNoteOnVol, 0x0);
-        ASSERT_EQ(seq.field_16_running_status, 0x90);
-        ASSERT_EQ(seq.field_90_delta_value, DvConv(0x2));
-    }
-
-    // Note on test
-    {
-        ClearTestData();
-
-        SeqStruct seq = {};
-        _ss_score[0] = &seq;
-        seq.field_16_running_status = 0x90;
-        seq.field_17_channel_idx = 5;
-
-        unsigned char midiData[] = 
-        {
-            0xA, // Note
-            0x22, // Velocity (on)
-            0x2, // Delta value
-        };
-        seq.field_0_seq_ptr = midiData;
-
-        _SsGetSeqData(0, 0);
-
-        ASSERT_EQ(seq.field_17_channel_idx, 5);
-        ASSERT_EQ(gMidiTest.mNoteOnCalls, 1);
-        ASSERT_EQ(gMidiTest.mNoteOnNote, 0xA);
-        ASSERT_EQ(gMidiTest.mNoteOnVol, 0x22);
-        ASSERT_EQ(seq.field_16_running_status, 0x90);
-        ASSERT_EQ(seq.field_90_delta_value, DvConv(0x2));
-    }
-
-    // Running status Note on test
-    {
-        ClearTestData();
-
-        SeqStruct seq = {};
-        _ss_score[0] = &seq;
-
-        unsigned char midiData[] = 
-        {
-            0x95, // Note on, channel 5
-            0xA, // Note
-            0x22, // Velocity (on)
-            0x2, // Delta value
-        };
-        seq.field_0_seq_ptr = midiData;
-
-        _SsGetSeqData(0, 0);
-
-        ASSERT_EQ(seq.field_17_channel_idx, 5);
-        ASSERT_EQ(gMidiTest.mNoteOnCalls, 1);
-        ASSERT_EQ(gMidiTest.mNoteOnNote, 0xA);
-        ASSERT_EQ(gMidiTest.mNoteOnVol, 0x22);
-        ASSERT_EQ(seq.field_16_running_status, 0x90);
-        ASSERT_EQ(seq.field_90_delta_value, DvConv(0x2));
-    }
-
-    // Program change test
-    {
-        ClearTestData();
-
-        SeqStruct seq = {};
-        _ss_score[0] = &seq;
-
-        unsigned char midiData[] = 
-        {
-            0xC5, // Program change channel 5
-            0xA, // Change
-        };
-        seq.field_0_seq_ptr = midiData;
-
-        _SsGetSeqData(0, 0);
-
-        ASSERT_EQ(seq.field_17_channel_idx, 5);
-        ASSERT_EQ(gMidiTest.mProgramChangeCalls, 1);
-        ASSERT_EQ(gMidiTest.mProgramChangeValue, 0xA);
-        ASSERT_EQ(seq.field_16_running_status, 0xC0);
-        ASSERT_EQ(seq.field_90_delta_value, DvConv(0));
-    }
-
-    // Running status program change test
-    {
-        ClearTestData();
-
-        SeqStruct seq = {};
-        _ss_score[0] = &seq;
-        seq.field_16_running_status = 0xC0;
-        seq.field_17_channel_idx = 5;
-
-        unsigned char midiData[] = 
-        {
-            0xA, // Change
-        };
-        seq.field_0_seq_ptr = midiData;
-
-        _SsGetSeqData(0, 0);
-
-        ASSERT_EQ(seq.field_17_channel_idx, 5);
-        ASSERT_EQ(gMidiTest.mProgramChangeCalls, 1);
-        ASSERT_EQ(gMidiTest.mProgramChangeValue, 0xA);
-        ASSERT_EQ(seq.field_16_running_status, 0xC0);
-        ASSERT_EQ(seq.field_90_delta_value, DvConv(0));
-    }
-
-    // Control change test
-    {
-        ClearTestData();
-
-        SeqStruct seq = {};
-        _ss_score[0] = &seq;
-
-        unsigned char midiData[] = 
-        {
-            0xB5, // Control change channel 5
-            0xA, // Change
-        };
-        seq.field_0_seq_ptr = midiData;
-
-        _SsGetSeqData(0, 0);
-
-        ASSERT_EQ(seq.field_17_channel_idx, 5);
-        ASSERT_EQ(gMidiTest.mControlChangeCalls, 1);
-        ASSERT_EQ(gMidiTest.mControlChangeValue, 0xA);
-        ASSERT_EQ(seq.field_16_running_status, 0xB0);
-        ASSERT_EQ(seq.field_90_delta_value, DvConv(0));
-    }
-
-    // Running status Control change test
-    {
-        ClearTestData();
-
-        SeqStruct seq = {};
-        _ss_score[0] = &seq;
-        seq.field_16_running_status = 0xB0;
-        seq.field_17_channel_idx = 5;
-
-        unsigned char midiData[] = 
-        {
-            0xA, // Change
-        };
-        seq.field_0_seq_ptr = midiData;
-
-        _SsGetSeqData(0, 0);
-
-        ASSERT_EQ(seq.field_17_channel_idx, 5);
-        ASSERT_EQ(gMidiTest.mControlChangeCalls, 1);
-        ASSERT_EQ(gMidiTest.mControlChangeValue, 0xA);
-        ASSERT_EQ(seq.field_16_running_status, 0xB0);
-        ASSERT_EQ(seq.field_90_delta_value, DvConv(0));
-    }
-
-    // TODO: Meta event test
-    {
-
-    }
-
-    // TODO: Running status Meta event test
-    {
-
-    }
-
-    // TODO: EOF meta event test
-    {
-
-    }
-
-    // TODO: Running status EOF meta event test
-    {
-
-    }
-
-    // Pitch bend test
-    {
-        ClearTestData();
-
-        SeqStruct seq = {};
-        _ss_score[0] = &seq;
-
-        unsigned char midiData[] = 
-        {
-            0xE5, // Pitch bend channel 5
-            // Bend value read in a helper
-        };
-        seq.field_0_seq_ptr = midiData;
-
-        _SsGetSeqData(0, 0);
-
-        ASSERT_EQ(seq.field_17_channel_idx, 5);
-        ASSERT_EQ(gMidiTest.mPitchBendCalls, 1);
-        ASSERT_EQ(seq.field_16_running_status, 0xE0);
-        ASSERT_EQ(seq.field_90_delta_value, DvConv(0));
-    }
-
-    // Running status Pitch bend test
-    {
-        ClearTestData();
-
-        SeqStruct seq = {};
-        _ss_score[0] = &seq;
-        seq.field_16_running_status = 0xE0;
-        seq.field_17_channel_idx = 5;
-
-        unsigned char midiData[] = 
-        {
-            0x0, // Nothing
-        };
-        seq.field_0_seq_ptr = midiData;
-
-        _SsGetSeqData(0, 0);
-
-        ASSERT_EQ(seq.field_17_channel_idx, 5);
-        ASSERT_EQ(gMidiTest.mPitchBendCalls, 1);
-        ASSERT_EQ(seq.field_16_running_status, 0xE0);
-        ASSERT_EQ(seq.field_90_delta_value, DvConv(0));
-    }
-
-    // TODO: EOF test
-    {
-
-    }
-}
-
-static void Test_SsPitchFromNote()
-{
-    for (int note = 0; note < 127; note++)
-    {
-        for (int fine = 0; fine < 127; fine++)
-        {
-            for (int centre = 0; centre < 127; centre++)
-            {
-                for (int shift = 0; shift < 127; shift++)
-                {
-                    const short realRet = SsPitchFromNote(note, fine, centre, shift);
-                    const short ourRet = Our_SsPitchFromNote(note, fine, centre, shift);
-                    if (ourRet != realRet)
-                    {
-                        printf("Failed to match result on %d %d %d %d\n", note, fine, centre, shift);
-                        ASSERT_EQ(ourRet, realRet);
-                    }
-                }
-            }
-        }
-    }
-}
-
-void DoTests()
-{
-    printf("Tests start\n");
-    Test_SsPitchFromNote();
-    Test_SsVmSelectToneAndVag();
-    Test_SsVmDoAllocate();
-    Test_SsGetSeqData();
-    printf("Tests end\n");
-    PcsxReduxExit(0);
-}
