+++ psx_seq_player/test.hpp
@@ -1,0 +1,4 @@
+#ifndef TEST_HPP
+#define TEST_HPP
+void run_unit_tests();
+#endif
