+++ psx_seq_player/types.h
@@ -1,0 +1,26 @@
+#ifndef TYPES_H
+#define TYPES_H
+
+typedef signed char s8;
+typedef signed short s16;
+typedef signed int s32;
+typedef signed long long s64;
+typedef unsigned char u8;
+typedef unsigned short u16;
+typedef unsigned int u32;
+typedef unsigned long long u64;
+
+// #ifndef PSX
+// #include <inttypes.h>
+// #else
+// typedef char s8;
+// typedef short int16_t;
+// typedef int int32_t;
+// typedef long long int64_t;
+// typedef unsigned char uint8_t;
+// typedef unsigned short uint16_t;
+// typedef unsigned int uint32_t;
+// typedef unsigned long long uint64_t;
+// #endif
+
+#endif
